requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
redis==5.0.3
celery==5.3.6
psycopg2-binary==2.9.9

pyannote.audio==3.1.1
speechbrain==1.0.0
openai-whisper==20231117
torch
torchvision
torchaudio
